% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosh_phonx.r
\name{azChar}
\alias{azChar}
\title{a cute convert-to-string wrapper}
\usage{
azChar(x)
}
\arguments{
\item{x}{an arbitrary object to be converted}
}
\description{
tha azChar phonk takes a obj + turns it into a character obj + gives it baqq
}
\examples{
azChar()
}
\keyword{character,}
\keyword{convert}
\keyword{string,}

